# services/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Báscula API
  version: "1.0.0"
  description: |
    Multi-tenant farm labour and payroll API. One binary, one Postgres, four
    roles. Isolation is row level security keyed on `app.farm_id`, set from the
    access token inside the transaction that serves the request.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /v1/activities:
    get:
      tags: [activities]
      operationId: listActivities
      summary: The farm's activity catalogue
      description: |
        Same route, different projection: the weigher's list arrives without a
        single rate in it — `rate` is absent, not null. The RLS policies on the
        `activity_pay_*` tables say the same thing a second time, so a
        forgotten projection here still cannot leak a price.

        `on` asks which rate was in force on a given day, which is how a past
        payment stays explainable.
      x-action: activities.read
      x-roles: [owner, admin, weigher]
      parameters:
        - $ref: "#/components/parameters/QueryQ"
        - $ref: "#/components/parameters/QueryStatus"
        - name: category
          in: query
          description: Category id, or its exact name.
          schema: { type: string }
        - name: on
          in: query
          description: Which day's rate to report. Defaults to today.
          schema: { type: string, format: date }
      responses:
        "200":
          description: Grouped by category, then by name.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Activity" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [activities]
      operationId: createActivity
      summary: Register an activity
      description: |
        The pay scheme decides the shape. `unidad_trabajo` needs a `unitId` and
        nothing else may have one. `weekly_price` is only available to
        `unidad_trabajo`, and `explicit` is not an activity's rate source at
        all — it belongs to a single work record.

        The first rate period is created in the same transaction; a deferred
        constraint refuses an activity that ends the transaction without one.
      x-action: activities.write
      x-roles: [owner, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ActivityInput" }
      responses:
        "200":
          description: That id already existed here.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Activity" }
        "201":
          description: Registered, with its first rate period.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Activity" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: DUPLICATE_NAME.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/activities/{id}:
    parameters:
      - $ref: "#/components/parameters/PathID"
    patch:
      tags: [activities]
      operationId: updateActivity
      summary: Rename, recategorise, archive or unarchive an activity
      description: |
        `payScheme` and `rateSource` cannot be changed and sending either is a
        400. Work records already written are pinned to (activity_id,
        pay_scheme) by a composite foreign key, and their price shape was
        decided by that scheme on the day they were written; turning "tala por
        jornal" into a per-kilo activity would rewrite the meaning of money
        already earned. An activity that pays differently is a different
        activity.

        The price is not editable here either: it has its own route, because a
        new price opens a validity period rather than overwriting the old one.
      x-action: activities.write
      x-roles: [owner, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ActivityPatch" }
      responses:
        "200":
          description: The activity as it now stands.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Activity" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: DUPLICATE_NAME.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      tags: [activities]
      operationId: archiveActivity
      summary: Archive an activity
      description: |
        Logical. Its rate history is not deleted either: a payment made under
        an old price stays explainable.
      x-action: activities.write
      x-roles: [owner, admin]
      responses:
        "204": { description: Archived. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/activities/{id}/rates:
    parameters:
      - $ref: "#/components/parameters/PathID"
    get:
      tags: [activities]
      operationId: listActivityRates
      summary: The whole price history of one activity
      description: |
        Newest first. This is what makes a past payment explainable: the answer
        to "why was I paid this" is a row with a date on it.
      x-action: activities.rate.write
      x-roles: [owner]
      responses:
        "200":
          description: Every validity period.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/ActivityRate" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/activities/{id}/rate:
    parameters:
      - $ref: "#/components/parameters/PathID"
    put:
      tags: [activities]
      operationId: setActivityRate
      summary: Open a new price period
      description: |
        Decision 4: activity prices have a history with validity dates, exactly
        like the weekly collection price already did. This never edits an
        earlier period — a rate already frozen onto a work record has to stay
        explainable — except when `validFrom` names a period that already
        starts on that day, which is a correction rather than a new price.

        The consequence a client must know: a work record whose price is
        derived from a date has to be a single day. A wage from Tuesday to
        Tuesday has no single validity period.
      x-action: activities.rate.write
      x-roles: [owner]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RateInput" }
      responses:
        "200":
          description: The activity's price history, newest first.
          content:
            application/json:
              schema:
                type: object
                required: [activityId, rates]
                properties:
                  activityId: { type: string, format: uuid }
                  rates:
                    type: array
                    items: { $ref: "#/components/schemas/ActivityRate" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------------------------------------------------------------------
  # Catalogues
  # ---------------------------------------------------------------------------
...
  /v1/catalogs/work-units:
    get:
      tags: [catalogs]
      operationId: listWorkUnits
      summary: Units of work
      description: |
        Kilos, arrobas, canastas. `kgFactor` is what a unit weighs, for the
        farms that count in baskets and report in kilos.
      x-action: catalogs.read
      x-roles: [owner, admin, weigher]
      responses:
        "200":
          description: The farm's units.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/WorkUnit" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [catalogs]
      operationId: createWorkUnit
      summary: Add a unit of work
      description: Idempotent by (farm_id, lower(code)); answers 200.
      x-action: catalogs.write
      x-roles: [owner, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkUnitInput" }
      responses:
        "200":
          description: The row that code means on this farm.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkUnit" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }

  # ---------------------------------------------------------------------------
  # Work records — "labores"
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A 15-minute access token with claims `sub`, `farm_id`, `role`,
        `device_id`. The farm comes from here and from nowhere else.
  parameters:
    PathID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    QueryQ:
      name: q
      in: query
      description: Free-text needle, case-insensitive substring.
      schema: { type: string }
    QueryStatus:
      name: status
      in: query
      description: |
        `active` (the default), `inactive` for what has been logically removed,
        `all` for both.
      schema: { type: string, enum: [active, inactive, all] }
...
  responses:
    BadRequest:
      description: BAD_REQUEST — malformed body, unknown field, or a bad value.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: UNAUTHORIZED — no token, or a token this server did not issue.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: FORBIDDEN — the permission table refused this role.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: |
        NOT_FOUND. Also what a resource of another farm answers, and what an id
        that is not a UUID answers: nothing with that id exists here.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    ErrorCode:
      type: string
      description: |
        The machine-readable half of every failure. The client branches on
        this; the translation lives in the client, which is why `message` is
        English.

        This enum is compared against domain.AllCodes() by the contract test,
        so a code cannot exist in the server without appearing here.
      enum:
        - BAD_REQUEST
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - INTERNAL
        - TENANT_NOT_SET
        - INVALID_CREDENTIALS
        - EMAIL_NOT_VERIFIED
        - EMAIL_TAKEN
        - TOKEN_EXPIRED
        - TOKEN_REUSED
        - RATE_LIMITED
        - FARM_LIMIT_REACHED
        - FARM_SUSPENDED
        - WORK_RECORD_SETTLED
        - PAYABLE_ALREADY_CLAIMED
        - SETTLEMENT_ALREADY_VOID
        - ALREADY_REVERSED
        - NOTHING_TO_SETTLE
        - AMOUNT_EXCEEDS_BALANCE
        - INVALID_GEOMETRY
        - PLOT_HAS_ACTIVE_CROPS
        - NO_RATE_IN_FORCE
        - RANGE_NEEDS_FROZEN_RATE
        - DUPLICATE_DOCUMENT
        - DUPLICATE_NAME
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message:
              type: string
              description: English, for a developer and a log. Never shown to a picker.
            details:
              type: object
              additionalProperties: true
              description: |
                What a client needs to recover. The canonical case is
                PAYABLE_ALREADY_CLAIMED, where `winningSettlement` lets the
                phone re-derive its state.
...
    PayScheme:
      type: string
      description: |
        How an activity pays. Spanish values because they are the values stored
        in the Postgres enum, which the design documents fixed.
      enum: [contrato, tiempo, unidad_trabajo]
    RateSource:
      type: string
      description: |
        When the price froze. `explicit` — the caller named it, frozen at write
        time. `activity_dated` — the activity's rate in force on the record's
        day, frozen at write time. `weekly_price` — read from the week at
        settlement time, which is what the phone does today. The two derived
        forms require a single-day record.
      enum: [explicit, activity_dated, weekly_price]
...
    TimeUnit:
      type: string
      enum: [jornal, semanal, quincenal, mensual, personalizado]

    # -- auth ----------------------------------------------------------------
...
    WorkUnit:
      type: object
      required: [id, code, label]
      properties:
        id: { type: string, format: uuid }
        code: { type: string, example: kg }
        label: { type: string, example: Kilo }
        kgFactor:
          type: number
          format: double
          nullable: true
          description: What one of these weighs, for farms that count in baskets.
    WorkUnitInput:
      type: object
      required: [code]
      properties:
        id: { type: string, format: uuid }
        code: { type: string, minLength: 1 }
        label: { type: string, description: Defaults to `code`. }
        kgFactor: { type: number, format: double, nullable: true }
    ActivityRate:
      type: object
      description: |
        One validity period of an activity's price. Decision 4: the rate is not
        a loose column, it is a row with a `validFrom`, and a work record
        freezes the one in force on its day.
      required: [validFrom, rateCents]
      properties:
        validFrom: { type: string, format: date-time }
        rateCents: { type: integer, format: int64 }
        timeUnit:
          allOf: [{ $ref: "#/components/schemas/TimeUnit" }]
          nullable: true
        customQty: { type: number, format: double, nullable: true }
        customUnit: { type: string, nullable: true }
    Activity:
      type: object
      required: [id, name, categoryId, category, payScheme, rateSource]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        categoryId: { type: string, format: uuid }
        category: { type: string }
        payScheme: { $ref: "#/components/schemas/PayScheme" }
        rateSource: { $ref: "#/components/schemas/RateSource" }
        unitId:
          type: string
          format: uuid
          nullable: true
          description: |
            On the activity and not on its rate: the unit is what is being
            counted, not what it costs. That is what lets a weigher see he is
            weighing kilos while being unable to read a single price.
        archivedAt: { type: string, format: date-time, nullable: true }
        rate:
          allOf: [{ $ref: "#/components/schemas/ActivityRate" }]
          nullable: true
          description: |
            The period in force on the requested day. ABSENT — not null — for
            the weigher.
    ActivityInput:
      type: object
      required: [name, payScheme, rate]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1 }
        categoryId: { type: string, format: uuid, description: One of this or `category`. }
        category: { type: string, description: A name; created in the catalogue if new. }
        payScheme: { $ref: "#/components/schemas/PayScheme" }
        rateSource:
          allOf: [{ $ref: "#/components/schemas/RateSource" }]
          description: |
            Defaults to `activity_dated`. `weekly_price` only for
            `unidad_trabajo`; `explicit` is rejected — it belongs to a single
            work record, not to an activity.
        unitId:
          type: string
          format: uuid
          nullable: true
          description: Required for `unidad_trabajo`, forbidden otherwise.
        rate: { $ref: "#/components/schemas/RateInput" }
    ActivityPatch:
      type: object
      description: |
        `payScheme` and `rateSource` are listed so the failure is explicit:
        sending either is a 400, never a silent no-op.
      properties:
        name: { type: string }
        categoryId: { type: string, format: uuid }
        category: { type: string }
        status: { type: string, enum: [active, inactive] }
        payScheme: { type: string, description: Rejected. }
        rateSource: { type: string, description: Rejected. }
    RateInput:
      type: object
      required: [rateCents]
      properties:
        rateCents: { type: integer, format: int64, minimum: 1 }
        validFrom:
          type: string
          format: date
          description: Defaults to today. A period that already starts on this day is corrected.
        timeUnit:
          allOf: [{ $ref: "#/components/schemas/TimeUnit" }]
          nullable: true
        customQty: { type: number, format: double, nullable: true }
        customUnit: { type: string, nullable: true }

    # -- work records --------------------------------------------------------
...